+++ src/index.js
@@ -1,16 +1,17 @@
 import '../styles/main.scss';
 
-export * from './cards/index.js';
 export * from './icons/index.js';
-export * from './spinner/index.js';
 export * from './alerts/index.js';
+export * from './cards/index.js';
 export * from './confirm-dialog/index.js';
 export * from './dropdown-container/index.js';
 export * from './dropdown-menu/index.js';
 export * from './dropdown-multi-select/index.js';
-export * from './split-pane/index.js';
-export * from './searchbar/index.js';
+export * from './floating-select/index.js';
 export * from './responsive-navbar/index.js';
+export * from './searchbar/index.js';
+export * from './spinner/index.js';
+export * from './split-pane/index.js';
 export * from './layouts/application/index.js';
 export * from './extended-search/index.js';
 export * from './wizard/index.js';